Added explicit environment to role
This will ensure clarity when looking for the roles in AWS. While
cloudformation would prefix the environment, it would get truncated,
if the service name were too long.

Also bumped the boto3 version to ensure that some of the new
capabilities such as ECR image scanning work on all deployments.